+++ example/android/app/src/main/java/com/example/flunamaexample/MainActivity.java
@@ -9,7 +9,6 @@ public class MainActivity extends FlutterActivity {
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
-    Mapbox.getInstance(this, "pk.eyJ1IjoiZm9udGluYWxpcyIsImEiOiJjamo3N3QxZzkwcGZkM3Jzd3R0OHd3eG9jIn0.MDmSPZyfbQ7NP0toOzSnyg");
     GeneratedPluginRegistrant.registerWith(this);
   }
 }