+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3
+
+* Add the ability to set Mapbox access token in code
+
+## 0.0.2
+
+* Populate the README and add some more info about the package
+
 ## 0.0.1
 
 * Published package